corrected refactor in search component and changed export to default
